configure.py: introduce link-time optimization
This patch introduces link-time optimization (LTO) to the build.

The performance gains from LTO alone are modest (~7%), but it's vital ingredient
of effective profile-guided optimization, which will be introduced later.

In general, use of LTO is quite simple and transparent to build systems.
It is sufficient to add the -flto flag to compile and link steps, and use a
LTO-aware linker.
At compile time, -flto will cause the compiler to emit .o files containing not
machine code, but compiler's internal representation (IR). At link time, those
pieces of IR will be compiled together, allowing cross-object optimization.

Therefore compiling Scylla with LTO can be as simple as running configure.py
with --cflags=-flto. Unfortunately this will also cause all other
executables (like tests) to be linked with LTO. In other words, the entire
project will be largely recompiled from scratch for every test, and that's
impractical.

GCC has the ability to emit "fat" objects containing both LTO-ready IR and
machine code, which can be used for both LTO and non-LTO linking, which solves
the above problem. Unfortunately LLVM doesn't have such option and we have to
implement an alternative ourselves, which complicates the situation.

This patch settles for a dual compilation of all objects: from source to IR
files and from IR files to ELF objects. The rule for scylla then uses the IR
files for linking and rules for tests use the ELF files.

Due to it's high compile time cost, the optimization can be toggled with a
configure.py option. As of this patch, it's disabled by default.