install-dependencies: add packages llvm, moreutils, iproute
Package llvm (15 MiB) contains the llvm-profdata utility, which is necessary
for preparing LLVM profile files for use in PGO builds.

iproute is a 1.9 MiB package which provides essential network management tools
such as `ip`. We are interested in it for `ss`, which will be used as a minor,
optional helper in PGO scripts (to validate that addresses are free before
attempting to launch a cluster).

moreutils is a small (200 kiB) package of miscellaneous utilities.
We want it for `ts` -- an utility which timestamps its lines of standard
input. We will use it to timestamp the logs of PGO training scripts.

[avi: regenerate frozen toolchain]