symmetric_key: unload providers when exit
ASan detects the leak due to the memory allocated by
`OSSL_PROVIDER_load()`. this function gets called whe initializing
the global static variable of `inited`, but the allocated memory
never gets freed, hence the leak. but since the allocation only
happens when scylla starts, so this should not cause infinitely
increasing leak. but it still annoying if ASan is enabled when
the scylla executable is compiled.

in this change, a helper struct of `openssl_env` is defined to
unload the previously loaded providers in the struct's dtor. this
should silence the warning from ASan.

Fixes #2749
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>