doc: remove OSS-to-OSS upgrade guides
This commit removes the Open Source-to-Open Source upgrade
guides from the Enterprise documentation.

The files are not physically removed, but they are excluded
from publishing ith the exclude patterns in the conf.py file
and removed from the toctree in the upgrade/index.rst file.

In addition, this commit replaces the relative paths with URLs
in the raft.rst file to prevent errors during publishing the docs.

Closes #2789