reader_concurrency_semaphore_group: use try_emplace in add_or_update()
unordered_map::emplace may construct an object and then immediately
destroy it if we're updating rather than adding. This violates
the promise that the name is unique. Currently it's not a problem,
but once reader_concurrency_semaphore registers metrics, it violates
metric uniqueness.

Use try_emplace, which does nothing if the key exists.

Closes #2832