audit: show real value in `error` column for direct queries
`error` column in audit log used to show almost always `False` because what it
was checking was only the access rights - not the real success of the
operation. This commit fixes the issue by logging also the success state of
`process_authorized_statement()`.

On this occasion I also unified the ordering of columns.

Fixes #572

Closes #2914