configure.py: don't rebuild Seastar on profile changes
During PGO builds, Seastar depends on the profile file.
But Seastar's build system doesn't give a way to specify
external profile files as dependencies, so the implementation of
this dependency was hacked into Scylla's build system, as a clear
of Seastar's build when the profile file was detected to be newer
than it.

But this hack is broken because there is more than one Seastar target
(libseastar.a and libseastar_testing.a). Because of it, Seastar's build
might be cleared by the build of libseastar_testing.a, even while
Scylla build processes are already using files produced by the build
of libseastar.a, pulling dependencies from under their feet for some
time.

This patch removes this hack altogether. After the patch, Seastar will
no longer be rebult when the profile changes. This should be OK because
profile changes are expected only to have a (minor) effect on performance,
and not correctness, and because it only affects incremental builds, not
clear builds.

The proper solution would be to add an option to Seastar's build which
allows specifying some external depenedency files.
A correct alternative which doesn't involve modifying Seastar might be
to invalidate the Seastar build while the profile file is *created*
(unpacked), rather than when Seastar targets are built.
But this solution is still hacky and might break down if some Seastar
targets which don't depend on the profile file are introduced.