test: fix encryption_at_rest test to work with raft
The test creates two cql test env objects within the same seastar_test
environment, which translates to reusing the data dir. The test,
however, was not re-using the host id between two executions.
This turned out to be critical to raft, which, on second
invocation, re-added a "new" host to the cluster and tried
to elect a leader in a two-node cluster where the first node
never came live.

Fixes #2995

Closes #3006