audit: fix partial mixup of fields in audit table update
A previous commit f3036652312 introduced a correct error logging
in audit, however, some mixup of the logging fields sneaked in.
There are more than one functions that logs audit information and uses
the same prepared statement. However, this commit only touched one of
them and in addition changed the field order in the prepared statement
but didn't change all of the places that use this prepared statement.
This commit sets the fields in the prepared statement back to their
original place and undo the change in the update function.

Ref #2998
Fixes #3109

Tests: unit tests
       audit dtests

This will specifically make the audit_test.TestCQLAudit.test_login dtest pass.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>

Closes #3081