doc: add the recommendation to upgrade drivers
This commit adds a requirement to upgrade ScyllaDB
drivers before upgrading ScyllaDB Enterprise.

The update is identical to the update in the Open Source
documentation added with this PR:
https://github.com/scylladb/scylladb/pull/14771#issuecomment-1643610509

I'm adding the information to the 2022.x-to-2023.1 upgrade guide
because the next upgrade guide will be based on this one, and
in this way, we can make sure the info will be propagated
with the upcoming releases.

Fixes https://github.com/scylladb/scylladb/pull/14771#issuecomment-1643610509