doc: replace the relative link to fix the warning
This commit updates the link to the Reverse Queries page.
The page was removed with https://github.com/scylladb/scylladb/pull/14190,
but the Enterprise 2019 upgrade guide contained
a reference to the removed page.
With this commit, the relative link is replaced with
the URL to the version where the Reverse Queries page still exists.