cql3: grammar: avoid using deprecated lw_shared_ptr assignment in batchStatementObjective
batchStatementObjective returns a lw_shared_ptr<something>, but assigns
a something directly to the pointer, which is now deprecated.

Avoid the deprecated operator and convert the object to a lw_shared_ptr
explicitly.

The reason to return a lw_shared_ptr is that antlr tries to use
the copy constructor of the original return value. Since the original
return value is std::unique_ptr, that fails. See 3b3761ce651747
("cql: restore properly setting raw query string for batch statements")
for details.

Closes #3228