kmip_host: Fix initial connection creation geting stuck
Fixes #3251

The re-using of connections is dependent on which failover
host is currently active. This break initial (verification)
check of hosts since we just connect to each host, without
activating the actual failover. And even if it was active,
we could have race conditions here that would cause at least
annoying starvation (sort of). Better to _not_ use the
hand-back+cache mechanism in the actual initial connect.

v2:
* Added comments/improved arg name