message/messaging_service: recognize OSS user tenant
OSS uses `statement:$user` as its isolation cookie to identify user
connections. In a mixed, OSS/Enteprise cluster, this is not recognized
by enterprise and it takes it as a yet unknown service level, creating it
on the spot. Furthermore, this service level seems to be
dropped/recreated periodically, for reasons I do not understand fully.
Avoid all this complications and the added confusion of seeing
`$user_read_concurrency_semaphore` in the logs, by hard-wiring a mapping
of `$user` to the default scheduling group.

Fixes: #3044

Closes #3281