doc: document KMS support as new in 2023.1.1
KMS support is added in the patch release 2023.1.1,
but we don't publish separate doc sets for patch releases.

This commit adds the note that support for KMS is only
available since the patch version 2023.1.1.

In addition, the note about when Encryption at Rest and KIMP
support were added has been removed - it's no longer
relevant, because the documentation is versioned.

Closes #3390