compaction: ICS: trigger tombstone GC if the gain outweighs the increased frequency
Today, tombstone GC is triggered if we find that a run satisfies the
tombstone threshold and interval params.

The interval was introduced as hack in C*, to prevent their single-
sstable tombstone GC logic from picking the same file over and over
again. The interval is checked against the write timestamp of the
sstable.

But that turns out to be suboptimal. We found that under high write
rate workloads, all tiers are potentially rewritten before the
interval expires, meaning that tombstone GC may never run.
Cross-tier logic is important for pushing upwards shadowed data
that was pushed to lower tiers as compaction output may be smaller
than input.

We can still apply interval as a staleness condition to reduce
write amplification, as reducing the frequency of tombstone GC
runs reduce the amount of times a byte will be rewritten.

Making the interval an actual frequency checker (as in TWCS'
expired_sstable_check_frequency_seconds) can be subptimal as the
default interval is 24h, and it might happen you check for tombstone
GC need now, but in a few hours, the threshold is met, and you miss
the opportunity to remove garbage. And that garbage will add to
the read amplification until the next opportunity to trigger GC.
Another reason is that interval serving as a staleness threshold
applied against each sstable run is lost.

Instead, we're going to use the interval as a condition to control
write amplification. If the interval is not satisfied, tombstone
GC will still be considered if the gain of removing garbage
outweights the increased frequency.
Previously, it could happen that a run is all garbage, but the
interval was preventing it from being picked, until regular
picked it or the interval itself expires, whichever comes first.

The default threshold is 0.2, meaning that a sstable with vastly
expiration times, could be rewritten 5 times over the course of
5 days (assuming interval of 1 day), causing a write amp of 3
(Sizeof(SSTable) * (1 + 0.8 + 0.4 + 0.6 + 0.2)). This is preserved
as threshold is only met after interval expires in this scenario.

But if the sstable have chunks of it with little difference in
expiration time, then interval condition will be used to increase
the threshold (to a minimum of 0.5), reducing the number of times
a sstable will be rewritten to a maximum of log2 N, adding only
one to the write ampl, as we'll be (potentially) halving the amount
of data on every GC round.

Fixes #3069.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>

Closes scylladb/scylla-enterprise#3415