ldap_connection: Check connection status before injecting await_result promise
When the await_result() is called, it checks if the ldap_search_ext()'s
return code is LDAP_SUCCESS. However, ldap can sometimes return back
success _after_ calling the connection::set_status() aborting it. The
latter method walks the registered promises resolving those with
exception, but the newcoming promise is not yet there, so it will remain
unresolved forever.

The fix is in making sure that the promise is injected while the status
is still up.

fixes: #3526

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>

Closes scylladb/scylla-enterprise#3528