ICS: Respect staleness condition for sstable runs possibly shadowing data in memtable
One of the conditions for GCing a tombstone is that its write timestamp
is older than memtable's min timestamp (used for calculating the
max purgeable timestamp).
It was found that ICS could submit sstables for tombstone compaction
despite none of the data can be GCed, due to overlapping in timestamp
with memtable. That's to avoid data resurrection.

This patch makes the file staleness (or tombstone interval) mandatory
for sstable runs which tombstones possibly shadow data in memtable,
and that is true if sstable's newest data is not older than
memtable's oldest data.
Shadowed data must be older than tombstone, so for memtables to have
older data than tombstones, out-of-order writes must have happened.
The staleness condition prevents ICS from doing useless attempts to
GC data in a sstable that was recently compacted, but bumped into
conditions that prevented most of the data (or even all of) from
being GCed.

Fixes #3561.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>

(contains part of af3bb1e8c664a4e082cbe9aa13319a8e8572b350)