workload prioritization: set unknown service levels shares to 1000
When a node restarts in a cluster with active workloads on service
levels that are different from the default, there is a time window in
which this node still haven't acquired the service level configuration
but rpc connections are opened on existing service levels. We have a
mechanism in place to mitigate this phenomena which speculatively
creates this service level with 1 share since when the real
configuration will be eventually acquired, this service level will be
either removed or updated to its real shares value.
However, since the number of shares also affects the static memory
partitioning between service levels it is going to receive ~1/1000 of
the memory which is too little even for simple read operations.
Here we change the default shares to be 1000. There are still problems
with this approach since we don't know what is the order of magnitude of
the other service levels, but a share of 1 is bad for almost every
reasonable configuration. So a value of 1000 should cover a wider range,
of configurations.

Fixes #3763

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>

Closes scylladb/scylla-enterprise#3842

(cherry picked from commit cc866ac6af0508888a70ef9bb578a7b3aa51a397)