doc: add Enterprise OS support to a separate file
This commit adds the information about Enterprise OS support
to an Enterprise-only file.
This update is necessary in order to split the Enterprise
and Open Source information while keeping the same page URL.

This commit is a follow-up to a corresponding update
in the scylladb/scylladb repository: https://github.com/scylladb/scylladb/pull/17310
When the above change is merged into the enterprise branch,
the file added with this commit will displayed on the OS support
page in the Enterprise docs.

Refs https://github.com/scylladb/scylladb/issues/15484

Closes scylladb/scylla-enterprise#3908

(cherry picked from commit f426c8b99aa63c2093e18cf4e537c1e1959efb83)