pull_github_pr.sh: update PR with `closes` prefix to avoid backporting wrong SHA commit
Today when we need to merge a PR with one commit, we add the `closes` prefix and then merge it using the `pull_github_pr.sh` script. This is causing a difference in commit sha between the one we see in the PR and the promoted commit. Causing our automation bot (Mergify) to backport wrong commit SHA.

This change will make sure that the `closes` prefix will also be updated in the PR.

Since this change requires us to use the `gh` tool (GitHub Cli), switching to use only this instead of `curl` for GitHub API

(cherry picked from commit c06e290d10e2205b27de7a6445b4f4377900aed4)