test_rpc_compression: test_dict_training: lower update period to fix flakiness
I forgot to adjust the dictionary polling period for this test,
and the default period is equal to the timeout of the test.
So sometimes nodes wouldn't update the dictionary within the timeout,
resulting in a flaky failure.

This patch fixes that.

(cherry picked from commit 719da493347b3944fcd985979ccbdcd6f3d77fe9)

Closes scylladb/scylla-enterprise#4465