db: config: add internode_compression_enable_advanced
Adds an extra flag, which can be used completely disables
the new RPC compression implementation, and falls back to the old one.

Using this flag, the new implementation is disabled by default,
to minimize the possibility of regressions in a patch release
for users who don't really want to use the new feature.