encryption: Fix encrypted components mask check in describe
Fixes scylladb/scylladb#22401

In the fix for scylladb/scylla-enterprise#892, the extraction and check for sstable component encryption mask was copied
to a subroutine for description purposes, but a very important 1 << <value> shift was somehow
left on the floor.

Without this, the check for whether we actually contain a component encrypted can be wholly
broken for some components.

Closes scylladb/scylladb#22398

(cherry picked from commit 7db14420b7b86a5e0ea5923d532b9b7e66fb3004)

Closes scylladb/scylla-enterprise#5230