test: Tune tls_test to make nodes wait each other on restart
The recently merged test restarts nodes one-by-one. After it it issues
the local_quorum select twards the cluster. The problem is that if
driver selects the last node in the row the query fails with unavailable
exception, because in 2024.1 nodes open CQL port _before_ they see each
other.

No idea why it happens in this branch, on master nodes start serving CQL
requests after they see each other, but maybe that's a coincidence.

Here it would be better to use manager.rolling_restart() facility, but
it's not backported, so do a lite version of it by hand.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>